// src/infrastructure/types/TypeUser.ts
export type TypeUser = {
    login: string;
    name?: string;

    id: number;

    avatar_url: string;

    html_url: string;

    followers_url: string;

    following_url: string;

    type: string;

    company?: string;

    blog?: string;

    email?: string;

    bio?: string;

    twitter_username?: string;

    public_repos: number;

    public_gists: number;

    followers: number;

    following: number;

    created_at: string;

    updated_at: string;
}